+++ src/game/scroll_targets.c
@@ -133,13 +133,13 @@ void patch_scroll_targets_interpolated(f32 delta) {
             Vtx* *verts = scroll->vertices;
             if (scroll->bhv < SCROLL_UV_X) {
                 u8 bhvIndex = MIN(scroll->bhv, 2);
-                for (u16 k = 0; k < scroll->size; k++) {
+                for (u32 k = 0; k < scroll->size; k++) {
                     f32 diff = wrap_f32(scroll->interpF32[k] - scroll->prevF32[k]);
                     verts[k]->n.ob[bhvIndex] = wrap_f32(scroll->prevF32[k] + diff * delta);
                 }
             } else {
                 u8 bhvIndex = MIN(scroll->bhv-SCROLL_UV_X, 1);
-                for (u16 k = 0; k < scroll->size; k++) {
+                for (u32 k = 0; k < scroll->size; k++) {
                     s32 diff = wrap_s32(scroll->interpS16[k] - scroll->prevS16[k]);
                     verts[k]->n.tc[bhvIndex] = wrap_s32(scroll->prevS16[k] + diff * delta);
                 }
